+++ database_ultimate_backup_lite/models/__init__.py
@@ -2,5 +2,6 @@
 
 from . import backup_provider_abstract
 from . import backup_provider_local
+from . import backup_provider_sftp
 from . import backup_config
 from . import backup_job